Logging restructuring, cleanup
Structured logging in a way that allows it to announce when a nested function starts and completes for better understanding of completion time. This outputs to both the log file and stdout

The logger also now automatically knows the folder name instead of having to manually input a new one for each project.

Got rid of an unused function.